Move Refund and RefundPayment to the shared folder, and delete the Refunds class since it's not needed anymore.
