+++ api-schema/src/main/java/org/stellar/anchor/api/platform/GetTransactionResponse.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.stellar.anchor.api.shared.Amount;
 import org.stellar.anchor.api.shared.Customers;
+import org.stellar.anchor.api.shared.Refund;
 import org.stellar.anchor.api.shared.StellarId;
 
 @Data
@@ -47,7 +48,7 @@ public class GetTransactionResponse {
   Instant transferReceivedAt;
 
   String message;
-  Refunds refunds;
+  Refund refunds;
 
   @SerializedName("stellar_transactions")
   List<StellarTransaction> stellarTransactions;
