+++ api-schema/src/main/java/org/stellar/anchor/api/shared/Refund.java
@@ -1,11 +1,10 @@
-package org.stellar.anchor.event.models;
+package org.stellar.anchor.api.shared;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.gson.annotations.SerializedName;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import org.stellar.anchor.api.shared.Amount;
 
 @Data
 @Builder
