+++ core/src/main/java/org/stellar/anchor/event/models/TransactionEvent.java
@@ -10,6 +10,7 @@ import lombok.Builder;
 import lombok.Data;
 import org.stellar.anchor.api.shared.Amount;
 import org.stellar.anchor.api.shared.Customers;
+import org.stellar.anchor.api.shared.Refund;
 import org.stellar.anchor.api.shared.StellarId;
 
 @Data
